refactor(421): extract decimal host family to vera/runtime/decimal.py
Step 8 of the codegen/api.py decomposition (#421).  Decimal (§9.6) is the one stateful family: it keeps a value-typed Python store (decimal_store: dict[int, PyDecimal]).  register_decimal(linker, ops_used, decimal_store, host_store_refs) takes both the store and the GC-introspection registry as parameters.  Because the shared host_decref_handle GC hook (kept inline in execute(), kind=3 path) must close over the same store, _decimal_store creation is lifted out of the Decimal branch to execute() top-level (an empty dict when Decimal is unused) -- the inline hook is otherwise untouched, minimising risk to the #706 destructor path.  api.py 1,735->1,480 lines (-66% from 4,358).  mypy + ruff clean; 68 decimal codegen + 72 gate/closure tests pass, and ch09_decimal/ch09_decimal_generics run end-to-end including under VERA_EAGER_GC=1 (decref-eviction stress).
